Added utils.py file
